+++ examples/src/index.js
@@ -9,6 +9,7 @@ import FilterExample from './js/FilterExample';
 import PreserveSelectOrderExample from './js/PreserveSelectOrderExample';
 import OptGroupExample from './js/OptGroupExample';
 import RestrictAvailable from './js/RestrictAvailableExample';
+import RtlExample from './js/RtlExample';
 
 ReactDOM.render(<BasicExample />, document.getElementById('basic-example'));
 ReactDOM.render(<OptGroupExample />, document.getElementById('optgroup-example'));
@@ -17,4 +18,5 @@ ReactDOM.render(<AlignTopExample />, document.getElementById('align-top-example'
 ReactDOM.render(<DisabledExample />, document.getElementById('disabled-example'));
 ReactDOM.render(<PreserveSelectOrderExample />, document.getElementById('preserve-select-order-example'));
 ReactDOM.render(<AllowDuplicatesExample />, document.getElementById('allow-duplicates-example'));
+ReactDOM.render(<RtlExample />, document.getElementById('rtl-example'));
 ReactDOM.render(<RestrictAvailable />, document.getElementById('restrict-available-example'));