+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New Features
 
+* [#76]: **(breaking)** Add `htmlDir` property to support RTL languages (defaults to LTR)
 * [#146]: Add styles for small viewport devices (<576px)
 * [#203]: Add `controlKey` to the `onChange` function such that developers may identify which control triggered the change
 * [#209]: **(breaking)** Add `iconsClass`, make icons more semantic, and support Unicode icons